query for current rest requests db correct tables

+++ src/main/java/by/bsuir/spp_project/dao/music/SongPostgreSQLDAO.java
@@ -11,13 +11,13 @@ import java.util.List;
 @Component("songDAO")
 @Service
 public class SongPostgreSQLDAO implements PostgreSQLDAO {
+
     @Autowired
     private SongRepository songRepository;
 
     @Override
     public void create(Object object) {
         Song song = (Song) object;
-
         songRepository.save(song);
     }
 
@@ -28,17 +28,34 @@ public class SongPostgreSQLDAO implements PostgreSQLDAO {
 
     @Override
     public Object readById(int id) {
-        return songRepository.getOne(id);
+        if (songRepository.existsById(id)) {
+            return (Song)songRepository.findById(id).get();
+        }
+        return null;
     }
 
     @Override
     public boolean update(Object object, int id) {
+        Song song = (Song) object;
+        if (songRepository.existsById(id)) {
+            song.setId(id);
+            songRepository.saveAndFlush(song);
+            return true;
+        }
         return false;
     }
 
     @Override
     public boolean delete(int id) {
-        songRepository.deleteById(id);
-        return true;
+        if (songRepository.existsById(id)) {
+            songRepository.deleteById(id);
+            return true;
+        }
+        return false;
+    }
+    @Override
+    public List getByValue(String column, Integer value) {
+
+        return null;
     }
 }
